KTOR-3677 Convert (X)ForwardedHeader to new API and rename plugin
