+++ ktor-server/ktor-server-test-suites/jvmAndNix/src/io/ktor/server/testing/suites/HttpServerCommonTestSuite.kt
@@ -14,7 +14,7 @@ import io.ktor.server.application.*
 import io.ktor.server.engine.*
 import io.ktor.server.http.*
 import io.ktor.server.plugins.autohead.*
-import io.ktor.server.plugins.forwardedsupport.*
+import io.ktor.server.plugins.forwardedheaders.*
 import io.ktor.server.plugins.statuspages.*
 import io.ktor.server.request.*
 import io.ktor.server.response.*
@@ -310,7 +310,7 @@ abstract class HttpServerCommonTestSuite<TEngine : ApplicationEngine, TConfigura
     @Test
     fun testProxyHeaders() {
         createAndStartServer {
-            application.install(XForwardedHeaderSupport)
+            application.install(XForwardedHeaders)
             get("/") {
                 call.respond(call.url { })
             }
