+++ ktor-server/ktor-server-tests/jvmAndNix/test/io/ktor/tests/server/plugins/HttpsRedirectPluginTest.kt
@@ -6,7 +6,7 @@ package io.ktor.tests.server.plugins
 
 import io.ktor.http.*
 import io.ktor.server.application.*
-import io.ktor.server.plugins.forwardedsupport.*
+import io.ktor.server.plugins.forwardedheaders.*
 import io.ktor.server.plugins.httpsredirect.*
 import io.ktor.server.response.*
 import io.ktor.server.routing.*
@@ -35,7 +35,7 @@ class HttpsRedirectPluginTest {
     @Test
     fun testRedirectHttps() {
         withTestApplication {
-            application.install(XForwardedHeaderSupport)
+            application.install(XForwardedHeaders)
             application.install(HttpsRedirect)
             application.routing {
                 get("/") {
